requirements.txt
apache-airflow
slack_sdk
asgiref
psycopg2-binary
snowflake-connector-python
snowflake-sqlalchemy
pandas
pyspark
pypika
tabulate
apache-airflow-providers-slack
apache-airflow-providers-snowflake
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
scikit-learn
streamlit
